Initial Commit
The main class for the Node has been created, as well a small demo/structured routine to demonstrate the data being passed for the different Node objects.

Slight issues in the beginning with immutable and mutable objects, i.e.  lists and dictionaries. This was solved using the .copy() function for dictionaries, and using a string to store the history of the nodes trajectory.

Tasks for next Commit:
- Fix get_duplicates() Method: Works for the HISTORY[x] key, but not MESSAGE??
- Small House Keeping