Remove Empty packets Method
Two methods create: One to discover the location of the empty packets within an objects buffer/(a node's buffer), Two: The ability to then remove the empty packets by referring to that list of empty packet locations.

Why so many methods? This will be simplified later. For now all the functionality is just being created.