Get Discovered connections Method Fix
Small bug found when running the get Discovered connections method; It would occasionally fail to read a newly discovered value... This was mainly due the order/sequence of events. A watch() function for testing has been added. This simulates a full run through of a network flood. As expected - It grows exponentially and takes a while to clean the network.

Way better than the previous attempt, which was using recursion, which quickly filled the machine stack.

Next Tasks:
- Research and begin implementing the API functionality