Congestion and Next Hop Functions complete
Will now return the congestion and status of each node, as well as the a function used to do a next hop, individual, or a full hop of each node in one command.

Next steps is to create one of the final functions which will start a thread and let the nodes run in the back ground. This should let the user of the API call the status of the node simulation while it's running in the background on the server.
This will include a start and stop function; The stop is self explanatory. However the Start function will have be invoked via a post command, with the ability to run a set number of iterations, or an infinite amount of iterations.